jdbc-to-jpa/src/main/java/net/shyshkin/study/jdbctojpa/jpa/PersonRepository.java
package net.shyshkin.study.jdbctojpa.jpa;

import net.shyshkin.study.jdbctojpa.domain.Person;
import org.springframework.stereotype.Repository;

import javax.persistence.EntityManager;
import javax.persistence.PersistenceContext;
import javax.persistence.TypedQuery;
import javax.transaction.Transactional;
import java.util.List;

@Repository
@Transactional
public class PersonRepository {

    @PersistenceContext
    EntityManager entityManager;

    public List<Person> findAllNamed() {

        TypedQuery<Person> namedQuery = entityManager.createNamedQuery("find_all_persons", Person.class);
        return namedQuery.getResultList();
    }

    public List<Person> findAll() {

        TypedQuery<Person> jpqlQuery = entityManager.createQuery("select p from Person p", Person.class);
        return jpqlQuery.getResultList();
    }

    public Person findById(int id) {
        return entityManager.find(Person.class, id);
    }

    public Person insertNew(Person person) {
        person.setId(null);
        return entityManager.merge(person);
    }

    public Person updatePerson(Person person) {
        return entityManager.merge(person);
    }

    public void deleteById(int id) {
        Person person = findById(id);
        entityManager.remove(person);
    }
}
